fix: remove unused zustand date store
